Cleanup and fixes for 64-bit port and toolchain update

+++ lib/runtime/comp/rqueue.h
@@ -18,7 +18,7 @@ typedef struct mbox rqueue;
 
 extern int rq_init(rqueue *rq, size_t size);
 extern void rq_close(rqueue *rq);
-extern ssize_t rq_receive(rqueue *rq, uint32_t *msg, size_t size);
-extern void rq_send(rqueue *rq, uint32_t *msg, size_t size);
+extern ssize_t rq_receive(rqueue *rq, uint64_t *msg, size_t size);
+extern void rq_send(rqueue *rq, uint64_t *msg, size_t size);
 
 #endif /* RQUEUE_H */
